Windows: fs getcacheparms
The get cache params output is supposed to include two values:

 . the size of the cache

 . the size of the cache in use

Windows no longer has a concept of an unused cache buffer.  All
buffers are inserted onto the freelist and are available for
recycling when the AFSCache file is created.  Instead of reporting
the used cache space as 0K, report it as the full cache in use.
It is likely to disturb users less.

Reviewed-on: http://gerrit.openafs.org/7284
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
(cherry picked from commit 7d0a0b66cca5dfd7f9c42032393632abadb1aa76)

Change-Id: I4ef3e27201505a18c1e841a7b10db43530fcfcf0
Reviewed-on: http://gerrit.openafs.org/7296
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>