Windows: handle KRB5KRB_AP_ERR_REPEAT errors
This patchset adds logic to handle a KRB5KRB_AP_ERR_REPEAT error
sent by a KDC for the Windows aklog and the afskfw library.  The
KRB5KRB_AP_ERR_REPEAT is sent when the KDC detects a replay.
However, the KDC replay detection is known to generate many false
positives.  If KRB5KRB_AP_ERR_REPEAT is received, wait one second
and retry the request.

This patchset does not apply to 'master' due to Kerberos
Compatibility SDK conversion.

Change-Id: I96518168d035f3c666761354348bd750718731a1
Reviewed-on: http://gerrit.openafs.org/8772
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>